requirements.txt
amqp==5.2.0
anyascii==0.3.2
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==3.2.4
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.2
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.8.0
boto3==1.34.158
botocore==1.34.158
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cramjam==2.8.3
cryptography==43.0.1
cssbeautifier==1.15.1
cssselect2==0.7.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
Django==5.0.8
django-anymail==10.3
django-ckeditor==6.7.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-js-asset==2.2.0
django-magicauth @ git+https://github.com/betagouv/django-magicauth.git@51f4709d8593016ac67428a87bbd2810070ae288
django-modelcluster==6.3
django-oauth-toolkit==2.4.0
django-otp==1.5.2
django-permissionedforms==0.1
django-simple-history==3.7.0
django-storages==1.14.4
django-taggit==5.0.1
django-treebeard==4.7.1
django-webpack-loader==3.1.0
django_csp==3.8
django_vite_plugin==3.0.3
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djlint==1.34.1
draftjs-exporter==5.0.0
drf-base64==2.0
drf-excel==2.4.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
EditorConfig==0.12.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
factory-boy==3.3.0
Faker==27.0.0
fastparquet==2024.5.0
filelock==3.15.4
filetype==1.2.0
freezegun==1.5.1
fsspec==2024.9.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html2text==2024.2.26
html5lib==1.1
identify==2.6.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kombu==5.4.2
l18n==2021.3
laces==0.1.1
lazy-object-proxy==1.10.0
lxml==5.2.2
mccabe==0.6.1
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
pillow_heif==0.18.0
pipdeptree==2.23.4
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
prompt-toolkit==3.0.47
psycopg2==2.9.9
py==1.11.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pylint==3.2.6
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pypdf==4.2.0
PyPDF3==1.0.6
pypng==0.20220715.0
pytest==8.0.0
pytest-django==4.8.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.4
referencing==0.35.1
regex==2023.12.25
reportlab==4.2.2
requests==2.32.3
requests-mock==1.12.1
rpds-py==0.20.0
s3transfer==0.10.2
sentry-sdk==2.10.0
sib-api-v3-sdk==7.6.0
six==1.16.0
soupsieve==2.5
sqlalchemy==2.0.30
sqlparse==0.5.0
svglib==1.5.1
telepath==0.3.1
text-unidecode==1.3
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.3
vine==5.1.0
virtualenv==20.26.5
wagtail==6.1.3
wcwidth==0.2.13
webencodings==0.5.1
Willow==1.8.0
wrapt==1.16.0
xhtml2pdf==0.2.16
